login User, modificaciones de los admins en los layouts

+++ app/helpers/sessions_helper.rb
@@ -1,8 +1,14 @@
 module SessionsHelper
 
-  def log_in(admin)
+  def log_in_admin(admin)
 
     session[:admin_id] = admin.id
+
+  end
+
+  def log_in_user(user)
+
+    session[:user_id] = user.id
 
   end
 
@@ -14,13 +20,29 @@ module SessionsHelper
     end
   end
 
-  def logged_in?
+  def current_user
+    if session[:user_id]
+      @current_user ||=
+      User.find_by(id: session[:user_id])
+    end
+  end
+
+  def logged_in_admin?
     !current_admin.nil?
   end
 
-  def log_out
-      session.delete(:admin_id)
-      @current_admin = nil
-    end
+  def logged_in_user?
+    !current_user.nil?
+  end
 
-end
+  def log_out
+      if !:admin_id.nil?
+        session.delete(:admin_id)
+        @current_admin = nil
+      end
+    else
+        session.delete(:user_id)
+        @current_user = nil
+      end
+      
+  end
